# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Nibook API
  version: 1.0.0
  description: Backend API for Nibook booking and business management platform
servers:
  - url: /api
    description: Base API path
paths:
...
  /profile/{id}:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get a profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/ProfileSuccess"
        "404":
          $ref: "#/components/responses/NotFoundError"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update a profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateRequest"
      responses:
        "200":
          $ref: "#/components/responses/ProfileSuccess"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /profile/by-slug/{slug}:
    get:
      operationId: getProfileBySlug
      tags: [profile]
      summary: Get a profile by its public slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/ProfileSuccess"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /analytics/{owner_id}:
    get:
      operationId: getAnalytics
      tags: [analytics]
      summary: Get business analytics
      security:
        - bearerAuth: []
      parameters:
        - name: owner_id
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          $ref: "#/components/responses/AnalyticsSuccess"
  /subscriptions/initialize:
    post:
      operationId: initializeSubscription
      tags: [subscriptions]
      summary: Initialize a Paystack subscription checkout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitializeSubscriptionRequest"
      responses:
        "200":
          $ref: "#/components/responses/SubscriptionInitializeSuccess"
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/ConfigurationError"
  /subscriptions/verify/{reference}:
    get:
      operationId: verifySubscription
      tags: [subscriptions]
      summary: Verify a Paystack payment reference
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/SubscriptionVerifySuccess"
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/ConfigurationError"
  /subscriptions/webhook:
    post:
      operationId: paystackWebhook
      tags: [subscriptions]
      summary: Paystack subscription event webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/WebhookReceived"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProfileUpdateRequest:
      type: object
      properties:
        business_name:
          type: string
        slug:
          type: string
        phone:
          type: string
        location:
          type: string
        bio:
          type: string
        category:
          type: string
        logo_url:
          type: string
        cover_url:
          type: string
        avatar_url:
          type: string
        mpesa_paybill:
          type: string
        mpesa_account:
          type: string
        whatsapp_enabled:
          type: boolean
        whatsapp_phone:
          type: string
        reminder_hours:
          type: integer
        cancellation_policy:
          type: string
        booking_widget_theme:
          type: string
        onboarding_completed:
          type: boolean
...
    InitializeSubscriptionRequest:
      type: object
      required:
        - email
        - plan_code
      properties:
        email:
          type: string
          format: email
        plan_code:
          type: string
        owner_id:
          type: string
...
  responses:
...
    ValidationError:
      description: Invalid request payload or missing fields
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...
    ConfigurationError:
      description: Server misconfiguration
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
              message:
                type: string